Site list based on IGBP and tags
This version adds the ability to set the site list to limit results by using IGBP. In addition, I added tags to the code block. A customize tag says that the block should be customized by the user. A comment with the word customize marks the area of code that the user would want to customize.